requirements.txt

fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
email-validator
pytest
